+++ src/collective/behavior/talcondition/utils.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 import logging
 from Products.CMFCore.Expression import Expression, createExprContext
-from Products.CMFCore.utils import getToolByName
+from plone import api
 
 logger = logging.getLogger('collective.behavior.talcondition')
 WRONG_TAL_CONDITION = "The TAL expression '%s' for element at '%s' is wrong.  Original exception : %s"
@@ -10,9 +10,7 @@ WRONG_TAL_CONDITION = "The TAL expression '%s' for element at '%s' is wrong.  Or
 def evaluateExpressionFor(obj, extra_expr_ctx={}):
     """Evaluate the expression stored in 'tal_condition' of given p_obj.
     """
-    res = True
-
-    # Check condition
+    # get tal_condition
     tal_condition = obj.tal_condition and obj.tal_condition.strip() or ''
     if not tal_condition:
         return True
@@ -21,12 +19,20 @@ def evaluateExpressionFor(obj, extra_expr_ctx={}):
 
     if hasattr(obj, 'context'):
         obj = obj.context
+    return _evaluateExpression(obj,
+                               expression=tal_condition,
+                               roles_bypassing_expression=roles_bypassing_talcondition,
+                               extra_expr_ctx=extra_expr_ctx)
 
-    member = getToolByName(obj, 'portal_membership').getAuthenticatedMember()
-    for role in roles_bypassing_talcondition:
+
+def _evaluateExpression(obj, expression, roles_bypassing_expression=[], extra_expr_ctx={}):
+    """Evaluate given p_expression extending expression context with p_extra_expr_ctx."""
+    res = True
+    member = api.user.get_current()
+    for role in roles_bypassing_expression:
         if member.has_role(str(role), obj):
             return res
-    portal = getToolByName(obj, 'portal_url').getPortalObject()
+    portal = api.portal.get()
     ctx = createExprContext(obj.aq_inner.aq_parent,
                             portal,
                             obj)
@@ -36,7 +42,7 @@ def evaluateExpressionFor(obj, extra_expr_ctx={}):
     for extra_key, extra_value in extra_expr_ctx.items():
         ctx.setGlobal(extra_key, extra_value)
     try:
-        res = Expression(tal_condition)(ctx)
+        res = Expression(expression)(ctx)
     except Exception, e:
         logger.warn(WRONG_TAL_CONDITION % (obj.tal_condition,
                                            obj.absolute_url(),
@@ -50,8 +56,8 @@ def applyExtender(portal, meta_types):
       We add some fields using archetypes.schemaextender to every given p_meta_types.
     """
     logger.info("Adding talcondition fields : updating the schema for meta_types %s" % ','.join(meta_types))
-    at_tool = getToolByName(portal, 'archetype_tool')
-    catalog = getToolByName(portal, 'portal_catalog')
+    at_tool = api.portal.get_tool('archetype_tool')
+    catalog = api.portal.get_tool('portal_catalog')
     catalog.ZopeFindAndApply(portal,
                              obj_metatypes=meta_types,
                              search_sub=True,